fix: [GSW-2040] CodeReview Feedback
