+++ packages/web/src/types/events.types.ts
@@ -1,0 +1,2 @@
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export type EventCallback = (...args: any[]) => void;
